+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- On error, the parser returns the position in the source code (used by MOGWAI STUDIO).
+
 ### Fixed
 
 ## [8.1.0] - 2026-02-25